+++ RELEASE_NOTES.md
@@ -10,8 +10,10 @@
  * Renamed types named RDBMSxxx to Rdbmsxxx to meet Go naming conventions
  * Streamlined [RdbmsClient](https://godoc.org/github.com/graniticio/granitic/rdbms) methods (xxxTagxxx methods removed, xxxParamxxx methods more flexible with the type of arguments they accept.)
  * [DatabaseProvider](https://godoc.org/github.com/graniticio/granitic/rdbms) interface no longer requires DatabaseFromContext - implementations can now implement the optional ContextAwareDatabaseProvider interface.
+ * [DatabaseProvider](https://godoc.org/github.com/graniticio/granitic/rdbms) interface no longer requires InsertIDFunc - implementations can now implement the optional NonStandardInsertProvider interface.
  * If RdbmsClient was created with a context, all operations on the underlying [sql.Db](https://golang.org/pkg/database/sql/#Db) object use the Context variants of methods where possible.
  * Transactions can now be opened with [sql.TxOptions](https://golang.org/pkg/database/sql/#TxOptions)
+ * Proxy interfaces introduced in front of [sql.Db](https://golang.org/pkg/database/sql/#Db) and [sql.TxOptions](https://golang.org/pkg/database/sql/#TxOptions) to facilitate testing.
  
 ### HttpServer